+++ compiler/tests-compiler/double-translation/direct_calls.ml
@@ -110,10 +110,13 @@ let%expect_test "direct calls with --effects=double-translation" =
                     === 2
                     ? f.cps.call(null, a0, a1)
                     : runtime.caml_call_gen_cps(f, [a0, a1])
-                  : a1
-                    ((f.l >= 0 ? f.l : f.l = f.length) === 1
-                      ? f(a0)
-                      : runtime.caml_call_gen(f, [a0]))
+                  : (f.l
+                      >= 0
+                      ? f.l
+                      : f.l = f.length)
+                    === 1
+                    ? a1(f(a0))
+                    : runtime.caml_call_gen_cps(f, [a0, a1])
                 : runtime.caml_trampoline_return(f, [a0, a1], 0);
        }
        function caml_exact_trampoline_cps_call(f, a0, a1){
@@ -131,10 +134,13 @@ let%expect_test "direct calls with --effects=double-translation" =
                     === 3
                     ? f.cps.call(null, a0, a1, a2)
                     : runtime.caml_call_gen_cps(f, [a0, a1, a2])
-                  : a2
-                    ((f.l >= 0 ? f.l : f.l = f.length) === 2
-                      ? f(a0, a1)
-                      : runtime.caml_call_gen(f, [a0, a1]))
+                  : (f.l
+                      >= 0
+                      ? f.l
+                      : f.l = f.length)
+                    === 2
+                    ? a2(f(a0, a1))
+                    : runtime.caml_call_gen_cps(f, [a0, a1, a2])
                 : runtime.caml_trampoline_return(f, [a0, a1, a2], 0);
        }
        function caml_exact_trampoline_cps_call$0(f, a0, a1, a2){